feat: 피처별 컴포넌트 및 로직 분리, useTimer 커스텀 훅 작성
